requirements.txt
pyspark>=3.3.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
python-dotenv>=0.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
